requirements.txt
numpy>=1.24.0
scipy>=1.10.0
pyyaml>=6.0
loguru>=0.7.0
fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart>=0.0.6
arize-phoenix>=4.3.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp>=1.20.0
nltk>=3.8.0
pymupdf4llm>=0.0.10
sentence-transformers>=2.2.0
langchain-text-splitters>=0.2.0
tiktoken>=0.4.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
requests>=2.31.0
ragas>=0.1.0# RAGAS & Evaluation
ragas
datasets>=2.14.0
langchain-openai>=0.0.5
langchain-ollama
langchain-huggingface
langchain-community

# Local Inference (HuggingFace)
torch
transformers
accelerate
bitsandbytes  # Optional optimization for loading models
